I have validated and enhanced the core engineering calculations and UI transparency of the pipeline analysis application.
Here's a summary of the key changes:
- Validated the FFS (B31G, Mod B31G, RSTRENG), growth correction (KNN-based), and failure simulation (dynamic clustering) modules, confirming their logical and computational soundness.
- Corrected the Folias factor formula in `calculate_b31g` to align with the published ASME B31G standard.
- Renamed `calculate_rstreng_effective_area_single` to the more descriptive `calculate_rstreng_level1` and updated its documentation and all call sites for clarity.
- Added explanatory expanders in the UI for both the Corrosion Assessment and Multi-Year Comparison views. These additions make the underlying assumptions of the FFS methods, growth correction, and failure simulation transparent to you, increasing confidence in the results.
